+++ src/YearCalendar.jsx
@@ -68,6 +68,14 @@ class YearCalendar extends React.Component {
       showDateInput: p.showDateInput,
       orient: ['top', 'left'],
       prefixCls: 'kuma-calendar',
+      // disabledDate: (current) => {
+      //   if (typeof p.disabledDate === 'function' && current) {
+      //     const date = current.clone();
+      //     date.getTime = current.valueOf;
+      //     return p.disabledDate(date);
+      //   }
+      //   return false;
+      // },
     };
     const pickerOptions = {
       disabled: p.disabled,
@@ -120,7 +128,10 @@ class YearCalendar extends React.Component {
               />
               {p.hasTrigger ? <Icon name="riqi" className={`kuma-calendar-trigger-icon ${showClear ? 'kuma-calendar-trigger-icon__has-clear' : ''}`} /> : null}
               {showClear
-                ? <i className="kuma-icon kuma-icon-close" onClick={this.clearValue} />
+                ? <i
+                  className="uxcore-icon uxicon-biaodanlei-tongyongqingchu kuma-icon-close"
+                  onClick={this.clearValue}
+                />
                 : null}
             </span>
           );